pyproject.toml
[tool.poetry]
name = "1chan"
version = "0.1.0"
description = ""
authors = ["Vladislav <barabayne@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
ruff = "^0.6.9"
pre-commit = "^4.0.1"
sqlalchemy = "^2.0.35"
fastapi = "^0.115.0"
pydantic-settings = "^2.5.2"
alembic = "^1.13.3"
pytest = "^8.3.3"
uvicorn = "^0.32.0"
asyncpg = "^0.30.0"
gunicorn = "^23.0.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
